requirements.txt
flask>=1.1.0
Flask-Bootstrap4>=4.0.0
requests>=2.22.0
